refact(modularization/settings): categorizing settings as module
