+++ dashboard/src/modules/settings/widgets/configuration.tsx
