+++ dashboard/src/modules/settings/widgets/index.ts
