+++ dashboard/src/routes/_dashboard/settings.lazy.tsx
@@ -1,6 +1,6 @@
 
 import { Page, VStack } from '@marzneshin/components'
-import { CertificateWidget } from '@marzneshin/features/settings';
+import { CertificateWidget } from '@marzneshin/modules/settings';
 import { SubscriptionSettingsWidget } from '@marzneshin/features/subscription-settings';
 import { createFileRoute } from '@tanstack/react-router';
 import { useTranslation } from 'react-i18next'
